Fix multiple C++ build errors: type mismatches, operator== mismatch, and id_generator array overflow

+++ src/xrGame/Level_network_messages.cpp
@@ -81,7 +81,7 @@ void CLevel::ClientReceive()
 		m_dwRPS += P->B.count;
 		//-----------------------------------------------------
 		u16 m_type;
-		u16 ID;
+		u32 ID;
 		P->r_begin(m_type);
 		switch (m_type)
 		{
